### 2018-07-04 #807 @whedon
[PRE REVIEW]: MaDaTS: Managing Data on Tiered Storage for Scientific Workflows
**Submitting author:** @dghoshal-lbl (<a href="http://orcid.org/0000-0002-6819-6949">Devarshi Ghoshal</a>)
**Repository:** <a href="https://github.com/dghoshal-lbl/madats" target ="_blank">https://github.com/dghoshal-lbl/madats</a>
**Version:** v1.1.2

Thanks for submitting your paper to JOSS dghoshal-lbl. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.

dghoshal-lbl if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
