### 2018-07-04 4ace695
affiliation fix for JOSS paper
