### 2018-07-09 #807 @arfon
@danielskatz - is this something you might be interested in editing?
