### 2018-07-09 #807 @danielskatz
If the authors are ok with me editing, I'm happy to.  (There's a small possibility of an informal conflict, though not a formal one according to the JOSS guidelines.)

@dghoshal-lbl - what do you think?
