### 2018-07-09 #807 @dghoshal-lbl
@danielskatz I am fine with you being the editor. However, I am not sure if there is a possibility of a conflict.
