### 2018-07-09 #807 @danielskatz
It's not a formal conflict, so let's not worry about it as long as you are ok with me doing this, which you are.

@dghoshal-lbl - can you suggest potential reviewers?   In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission.
