### 2018-07-09 #807 @dghoshal-lbl
@danielskatz I am not sure if I have a good idea of who can be the potential reviewers. Since the code is in Python, I think anyone who has a working knowledge of Python can review this. Last time, I submitted a Python software to JOSS, I think @krother was the reviewer.
