### 2018-07-23 #830 @whedon
[REVIEW]: MaDaTS: Managing Data on Tiered Storage for Scientific Workflows
**Submitting author:** @dghoshal-lbl (<a href="http://orcid.org/0000-0002-6819-6949">Devarshi Ghoshal</a>)
**Repository:** <a href="https://github.com/dghoshal-lbl/madats" target ="_blank">https://github.com/dghoshal-lbl/madats</a>
**Version:** v1.1.2

- [x] **Authorship:** Has the submitting author (dghoshal-lbl) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- [x] **Authorship:** Has the submitting author (dghoshal-lbl) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
