### 2018-07-23 #830 @danielskatz
👋  @ian-taylor & @gflofst - please go ahead and start your reviews

See the first comment in the issue for details on how to do the review, and  https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and https://joss.readthedocs.io/en/latest/review_criteria.html for more details. In brief, what I'm looking for is you both to check off the 18 items on your checklists.

If you have problems, feel free to comment here or email me.

Finally, a note that @ian-taylor will not be able to do his review until the 2nd week of August.
