### 2018-07-30 #830 @gflofst
The license is not a standard license, but a University of California/DOE license. It is very straightforward reminiscent of the MIT license.
