### 2018-07-30 #830 @danielskatz
Thanks @gflofst

👋 @dghoshal-lbl - Note that JOSS requires using [an OSI-approved license](https://opensource.org/licenses/alphabetical) - which is this?
