### 2018-07-30 #830 @gflofst
It has no text citing it as a particular license. Instead, it lists that it is approved by Uni. Cal. and LBL (and DOE). The license text is basically MIT (do what you want, cite this code, we are not responsible for what you do or what happens because of use).
