### 2018-07-30 #830 @dghoshal-lbl
@danielskatz @gflofst It is a modified BSD license.
