### 2018-07-30 #830 @gflofst
In the setup.py, it says BSD. How exacting should that match be to be acceptable?
