### 2018-07-30 #830 @danielskatz
Modified in any way is not acceptable.

To be published by JOSS, the code needs to have an OSI-approved license.
