### 2018-07-30 #830 @dghoshal-lbl
There is an additional paragraph at the end to the original BSD license (due to lab policies).
