### 2018-07-30 #830 @dghoshal-lbl
Is it mandatory to have the exact same license? Previously one of the papers was accepted because the license is still BSD, with an additional paragraph.
