### 2018-07-30 #830 @danielskatz
Yes, it is mandatory.  Can I check the process for the other paper?  Which was it?
