### 2018-07-30 #830 @gflofst
Community guidelines are either missing or not obvious
