### 2018-07-30 #830 @dghoshal-lbl
@danielskatz
i) This is the paper: https://joss.theoj.org/papers/f68f92a740c6e42365025be9a9efed1b
ii) Software package: https://zenodo.org/record/290299#.W1-isdhKjOQ (Please note license on the right side, which says --  BSD 3-Clause "New" or "Revised" License (BSD-3-Clause))
iii) And the review comments: https://bitbucket.org/dghoshal/frieda/issues/1/joss-review-comments
