### 2018-07-30 #830 @dghoshal-lbl
@gflofst A sample storage.yaml file is created when setup.py is run. Did the build process fail?
