### 2018-07-31 #830 @danielskatz
👋 @arfon  - please weigh in on the license discussion for this submission, as well as for https://doi.org/10.21105/joss.00164
