### 2018-07-31 #830 @arfon
> Is it mandatory to have the exact same license? Previously one of the papers was accepted because the license is still BSD, with an additional paragraph.

I think this statement is contradictory. The whole point of having a standard set of approved and vetted open source licenses is that they are used verbatim _without modification_.

This submission should be considered blocked by this issue and I think the JOSS editorial board needs to discuss the status of https://doi.org/10.21105/joss.00164 as the non-standard license is problematic and something we should have caught (and likely would going forward with this https://github.com/openjournals/whedon-api/pull/35).
