### 2018-07-31 #830 @gflofst
@dghoshal-lbl Here is the output when I try to run setup. It looks like I should have created a directory, but this isn't specified in the instructions. Please review and adjust the instructions accordingly.
[setup.txt](https://github.com/openjournals/joss-reviews/files/2246568/setup.txt)
