### 2018-07-31 #830 @dghoshal-lbl
@gflofst There is no need to create any additional directory. The config/ directory is already present in the top-level directory of the source tree.
Unfortunately I am unable to reproduce the error on my side. I need some more information.
- Is this where you have checked out the source (including setup.py): /Users/gflofst/build/joss/run?
- Have you manually configured MADATS_HOME environment variable?
