### 2018-07-31 #830 @gflofst
I did the clone into joss so the top level directory is
/Users/gflofst/build/joss/madats

I do not have MADATS_HOME set in this terminal window. While the instructions after the two steps say to set it up, it should be a step in the instructions in order too. When I do set it to the above directory, it works. Having to set it to generate the file that will set it doesn't make sense.
