### 2018-07-31 #830 @dghoshal-lbl
You do not need to set MADATS_HOME manually. I will fix the instructions in the readme if that is not clear enough. However, if you run setup from the top level directory "/Users/gflofst/build/joss/madats" (where setup.py is), it should create the required files.
I was wondering why the setup log shows "/Users/gflofst/build/joss/run" as MADATS_HOME. What is this joss/run/ directory, if code is in joss/madats/?
