### 2018-07-31 #830 @gflofst
code is in joss/madats and that is where I run setup.py from. The only way I can get it to work properly is if I set MADATS_HOME manually before running setup.py.
