### 2018-07-31 #830 @dghoshal-lbl
That shouldn't happen. Can you please open up a new terminal and let me know if "echo $MADATS_HOME" prints anything (without you manually setting it up)? If it doesn't print anything, can you please rerun "python setup.py install" and send me the output of the build process?

I tried in a couple of different environments and unable to reproduce the error. Can you please check if there is a "storage.yaml" file present in config/ directory at the end of the build process?
