### 2018-07-31 #830 @gflofst
Oddly, I closed my terminal so that I cleared out any env vars I set manually and cleaned the repo clone and now it works. I am in the same directory and everything.

I'll consider this pass now. Not sure what my problem was. If you could clarify the readme instructions for setup to just have the source MADATS_HOME as the third step instead of the next paragraph, that would help.

I am slightly concerned that this is bash centric, but that might be my old timer personality speaking. Bash pretty well has taken over, but it is not a 100% solution. A note to adjust for other shells would make it 100%, but I am not sure how necessary that is these days.
