### 2018-07-31 #830 @gflofst
At this point, the only two pieces of contention are the license (out of my hands) and the need for a "how to contribute" set of instructions. Everything else looks good.
