### 2018-07-31 #830 @dghoshal-lbl
Thanks @gflofst. Here are the action items for me to fix:
i) Fix the readme to have "source MADATS_HOME" as the third step.
ii) A set of instructions on "how to contribute".
iii) Since the bash-centric command is only for setting up MADATS_HOME, I will add a few lines in the readme to setup MADATS_HOME on other environments.

Let me know if there are any other changes that you think are necessary. I will fix these and let you know.

I hope the licensing issue will be resolved soon.
