### 2018-07-31 #830 @danielskatz
We have an editors call on Friday where the licensing issue will be discussed - I think it's going to be blocking problem, so @dghoshal-lbl, you might want to think about what you are able to do or will choose to do if so.
