### 2018-08-07 #830 @danielskatz
@gflofst & @ian-taylor - please put this on hold on now.  We cannot proceed with the current license, so either the license will be changed, or we will remove this from the JOSS queue.  We've asked the author to let us know what he can do in the next three weeks, with the understanding that employer policies may not permit a satisfactory resolution in that time - if so, the paper will likely be withdrawn until it meets the JOSS policies at which time it could be resubmitted, and we could use the reviewing work that has already been done as a starting point.
