### 2018-08-10 178a12e
updated readme and contributing file
