### 2018-08-14 #830 @dghoshal-lbl
@gflofst I have updated the repository with the changes you proposed. I have updated the readme and added a 'CONTRIBUTING' file. The license is also changed to OSI approved 3-clause BSD license.

Please let me know if everything looks okay and if it needs any more edits/modifications.
