### 2018-08-14 #830 @gflofst
@dghoshal-lbl , the checklist for the contributions looks for 3 things:
Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
I see the first one, but not the second 2.
