### 2018-08-15 #830 @dghoshal-lbl
@gflofst The issues can be submitted through Github issues for the repository. Do you want me to add the information in the "CONTRIBUTING" file?
