### 2018-08-15 #830 @gflofst
@dghoshal-lbl I agree it is a bit redundant since it is expected that if it is on github that you'll respond to anything reported there, but I think it would be best to list it in the file just to be clear. For me, I'd list to use that or if it proves unresponsive (more than 1 week before some response), to contact me via email. I don't think that is necessary for this journal checklist though.
