### 2018-08-21 #830 @ian-taylor
ok, here is what I did to start from scratch using python 2.7 and pip version.x 18:

git clone https://github.com/dghoshal-lbl/madats.git
cd madats
virtualenv madats
source madats/bin/activate
pip install --upgrade pip
(and pip -V
pip 18.0 from /Users/ian/Sites/madats/madats/madats/lib/python2.7/site-packages/pip (python 2.7)
)
pip install -r requirements.txt
python setup.py install
source MADATS_HOME
py.test tests/test_madats.py

And I got:

============================================= test session starts ==============================================
platform darwin -- Python 2.7.10, pytest-3.4.0, py-1.5.4, pluggy-0.6.0
rootdir: /Users/ian/Sites/madats/madats, inifile:
collected 0 items / 1 errors

==================================================== ERRORS ====================================================
____________________________________ ERROR collecting tests/test_madats.py _____________________________________
ImportError while importing test module '/Users/ian/Sites/madats/madats/tests/test_madats.py'.
Traceback:
tests/test_madats.py:19: in <module>
    import madats
madats/lib/python2.7/site-packages/MaDaTS-1.1.2-py2.7.egg/madats/__init__.py:1: in <module>
    from madats.core.vds import VirtualDataSpace, VirtualDataObject, Task
madats/lib/python2.7/site-packages/MaDaTS-1.1.2-py2.7.egg/madats/core/vds.py:24: in <module>
    from scandir import scandir
E   ImportError: No module named scandir
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Interrupted: 1 errors during collection !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
=========================================== 1 error in 0.29 seconds ============================================

-- So, should I assume scandir should be available?, did I do something else wrong? or should this be added to the install requirements.txt file and/or instructions?
