### 2018-08-21 #830 @danielskatz
I was hoping you would check the COI and code of conduct ones before you started on other things :)
