### 2018-08-21 #830 @dghoshal-lbl
@ian-taylor Not sure how I missed that. I will add scandir to requirements.txt
