+++ requirements.txt
@@ -1,3 +1,4 @@
 pyyaml==3.12
 six>=1.4.1
 pytest==3.4.0
+scandir==1.9.0
