### 2018-08-25 #830 @ian-taylor
Its quite hard to get going on this without creating your own workflow from scratch.  None of the examples in the examples directory work as is.  To get any of these to work, as stated in each example you need to "Update config/storage.yaml with the appropriate storage tiers and specify the datadir to make this example work"

Now, the madats_workflow.yaml is generic and consists of two tasks.  The description in the UserGuide  "MaDaTS Workflow Description" provides a list of what the options mean but there is no overview of how the workflow works.

For example, one would expect a way of defining dependencies between tasks if this is a workflow system. How can that be achieved? or is this more of a batch oriented system i.e. execute task1 then task2, etc ?

What would be useful is a fully operating example that a programmer can run to try this system out, even if it is simple.    Also, it would be helpful to provide 2 or 3 other functional workflows that help walk a programmer through the process of building simple batch flows and then more complex workflows.
