### 2018-08-27 #830 @dghoshal-lbl
@ian-taylor The example section in Readme lists the steps to create a workflow of your own. But I think you are asking for a script or a sample workflow that you can tweak to build your own workflow. Is that correct?
