### 2018-08-29 #830 @dghoshal-lbl
@ian-taylor I think I have a few examples in the source tree. I will check those and see if I can add a few lines in the readme/user-guide on how to use those and build your own workflows using the examples as a reference. Hopefully that will help.
