### 2018-09-03 #830 @dghoshal-lbl
@danielskatz @ian-taylor There are a few examples in the examples/ directory in the source tree. I am adding more information about these examples in the user guide so that users can use these as reference.
