### 2018-09-05 #830 @dghoshal-lbl
@ian-taylor I have added Chapter-3 in UserGuide.pdf to explain some examples of using MaDaTS. Let me know if it addresses what you asked for.
