### 2018-09-28 #830 @danielskatz
👋  @gflofst - what about the community guidelines checklist item? Can you check this off, or let us you what else you think is needed?
